Refactor build configuration and enhance Telegram bot service
- Updated `build.gradle.kts` to replace Logback with Log4j2 for improved logging capabilities and streamlined dependency management.
- Refactored `TelegramBotService` to simplify message handling and improve command execution flow.
- Enhanced `BeanConfig` to provide a custom `ObjectMapper` for better date handling and serialization.
- Updated `MongoConfig` to include custom converters for `OffsetDateTime` support.
- Removed deprecated `TelegramCommands` class to clean up the codebase and improve maintainability.
- Adjusted user ID handling in `JobSearchRepository` and related services to use `Long` instead of `Int` for better compatibility with user data.
- Improved error handling and user feedback in various services, enhancing overall user experience.

+++ core_service/src/main/kotlin/com/jobsalerts/core/repository/JobSearchRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository
 
 @Repository
 interface JobSearchRepository : MongoRepository<JobSearchOut, String> {
-    fun findByUserId(userId: Int): List<JobSearchOut>
+    fun findByUserId(userId: Long): List<JobSearchOut>
     fun findByIdAndUserId(id: String, userId: Int): JobSearchOut?
     fun deleteByIdAndUserId(id: String, userId: Int): Long
-} 
+}
